+++ components/DailyTodos.jsx
@@ -3,31 +3,58 @@ import { DateContext } from '@/contexts/DateContext';
 import { LoginContext } from '@/contexts/LoginContext';
 import useTodoStore from '@/contexts/TodoStore';
 import { useTodoAddMutation } from '@/hooks/useTodoMutations';
-import TODO_QUERY_KEY from '@/hooks/useTodoQuery';
+import {
+  default as TODO_QUERY_KEY,
+  default as useTodosQuery,
+} from '@/hooks/useTodoQuery';
+import { isTodoIncludedInTodayView } from '@/utils/dateUtils';
 import { useQueryClient } from '@tanstack/react-query';
 import { Input, List } from '@ui-kitten/components';
 import { LexoRank } from 'lexorank';
 import { Fragment, useContext, useEffect, useState } from 'react';
-import { View } from 'react-native';
+import { Text, View } from 'react-native';
 import DailyTodo from './DailyTodo';
 
 const DailyTodos = () => {
+  const queryClient = useQueryClient();
   const [input, setInput] = useState('');
   const { userId, accessToken } = useContext(LoginContext);
   const { selectedCategory } = useContext(CategoryContext);
-  const { mutate: addTodo, isSuccess } = useTodoAddMutation();
-  const { selectedDate, setSelectedDate } = useContext(DateContext);
-
-  const queryClient = useQueryClient();
+  const { mutate: addTodo, isSuccess: addTodoIsSuccess } = useTodoAddMutation();
+  const { selectedDate } = useContext(DateContext);
+  const {
+    isLoading,
+    error,
+    data,
+    isSuccess: isTodosQuerySuccess,
+  } = useTodosQuery(accessToken, userId);
+  const currentTodos = useTodoStore(state => state.currentTodos);
 
   useEffect(() => {
-    if (isSuccess) {
+    if (isTodosQuerySuccess) {
+      useTodoStore.setState({ todos: data });
+      let filteredTodos = data.filter(
+        todo =>
+          todo.categoryId === selectedCategory &&
+          isTodoIncludedInTodayView(
+            todo.startDate,
+            todo.endDate,
+            selectedDate.format('YYYY-MM-DD'),
+          ),
+      );
+      useTodoStore.setState({ currentTodos: filteredTodos });
+    }
+  }, [isTodosQuerySuccess, data, selectedCategory, selectedDate]);
+
+  useEffect(() => {
+    if (addTodoIsSuccess) {
       queryClient.invalidateQueries(TODO_QUERY_KEY);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isSuccess]);
+  }, [addTodoIsSuccess]);
 
-  const currentTodos = useTodoStore(state => state.currentTodos);
+  if (isLoading) return <Text>Loading...</Text>;
+  if (error) return <Text>Error: {error.message}</Text>;
 
   const renderTodo = ({ item, drag, isActive }) => {
     return (
